+++ src/app/locales/en/components/cluster-storage-add-form.js
@@ -1,6 +1,7 @@
 import generic from './cluster-storage-add-form/generic';
 import luma from './cluster-storage-add-form/luma';
 import ceph from './cluster-storage-add-form/ceph';
+import cephrados from './cluster-storage-add-form/cephrados';
 import posix from './cluster-storage-add-form/posix';
 import s3 from './cluster-storage-add-form/s3';
 import swift from './cluster-storage-add-form/swift';
@@ -11,6 +12,7 @@ export default {
   generic,
   luma,
   ceph,
+  cephrados,
   posix,
   s3,
   swift,